+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Comply with CRAN directive to remove references to packages not in
   depends/imports/suggests in tests (these were run optionally before).
+* Fix bugs in corner case handling when we hit `max.diffs`.
 
 ## v0.1.9